Implement access to player edit actions

+++ ReversiMvc/Controllers/PlayerController.cs
@@ -1,5 +1,6 @@
 ﻿#nullable disable
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using ReversiMvc.Services.Contracts;
 
@@ -11,12 +12,14 @@ public class PlayerController : Controller
     private readonly IPlayersRepository _repository;
     private readonly ILogger<PlayerController> _logger;
     private readonly string _currentUser;
+    private readonly UserManager<IdentityUser> _userManager;
 
-    public PlayerController(IPlayersRepository repository, ILogger<PlayerController> logger, ICurrentUserService currentUser)
+    public PlayerController(IPlayersRepository repository, ILogger<PlayerController> logger, ICurrentUserService currentUser, UserManager<IdentityUser> userManager)
     {
         this._repository = repository;
         this._logger = logger;
         this._currentUser = currentUser.Name;
+        this._userManager = userManager;
     }
 
     // GET: Players
@@ -45,31 +48,6 @@ public class PlayerController : Controller
         this._logger.LogInformation("User {User} has viewed the details of player {Player}", this._currentUser, playerEntity.Guid);
 
         return this.View(playerEntity);
-    }
-
-    // GET: Players/Create
-    public IActionResult Create()
-    {
-        return this.View();
-    }
-
-    // POST: Players/Create
-    // To protect from overposting attacks, enable the specific properties you want to bind to.
-    // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-    [HttpPost]
-    [ValidateAntiForgeryToken]
-    public async Task<IActionResult> Create([Bind("Guid,Name,Victories,Losses,Draws")] PlayerEntity playerEntity)
-    {
-        if (!this.ModelState.IsValid)
-        {
-            return this.View(playerEntity);
-        }
-
-        await this._repository.AddAsync(playerEntity);
-
-        this._logger.LogInformation("User {User} has created player {Player}", this._currentUser, playerEntity.Guid);
-
-        return this.RedirectToAction(nameof(this.Index));
     }
 
     // GET: Players/Edit/5
@@ -132,6 +110,11 @@ public class PlayerController : Controller
             return this.NotFound();
         }
 
+        if (id.Equals(this._currentUser))
+        {
+            return this.View("InvalidActionMessage", new InvalidActionViewModel { Message = "Je mag je eigen account niet verwijderen!" });
+        }
+
         var playerEntity = await this._repository.GetDbSet()
             .FirstOrDefaultAsync(m => m.Guid == id);
         if (playerEntity == null)
@@ -147,9 +130,27 @@ public class PlayerController : Controller
     [ValidateAntiForgeryToken]
     public async Task<IActionResult> DeleteConfirmed(string id)
     {
+        if (id.Equals(this._currentUser))
+        {
+            return this.View("InvalidActionMessage", new InvalidActionViewModel { Message = "Je mag je eigen account niet verwijderen!" });
+        }
+
         var playerEntity = await this._repository.GetDbSet().FindAsync(id);
 
         await this._repository.DeleteAsync(playerEntity);
+        
+        var user = await this._userManager.FindByIdAsync(id);
+        if (user != null)
+        {
+            var result = await this._userManager.DeleteAsync(user);
+            
+            if (!result.Succeeded)
+            {
+                throw new InvalidOperationException($"Unexpected error occurred deleting user.");
+            }
+        
+            this._logger.LogInformation("User with ID '{id}' deleted themselves.", id);
+        }
 
         this._logger.LogInformation("User {User} has deleted player {Player}", this._currentUser, playerEntity.Guid);
